Extract validation into module
