+++ r4/pom.xml
@@ -163,5 +163,10 @@
       <artifactId>spring-beans</artifactId>
       <scope>compile</scope>
     </dependency>
+    <dependency>
+      <groupId>gov.va.api.health</groupId>
+      <artifactId>validation</artifactId>
+      <version>${project.version}</version>
+    </dependency>
   </dependencies>
 </project>
