+++ r4/src/main/java/gov/va/api/health/r4/api/resources/Coverage.java
@@ -19,7 +19,7 @@ import gov.va.api.health.r4.api.elements.Extension;
 import gov.va.api.health.r4.api.elements.Meta;
 import gov.va.api.health.r4.api.elements.Narrative;
 import gov.va.api.health.r4.api.elements.Reference;
-import gov.va.api.health.r4.api.validation.ExactlyOneOf;
+import gov.va.api.health.validation.api.ExactlyOneOf;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.Valid;
