+++ r4/src/main/java/gov/va/api/health/r4/api/resources/CoverageEligibilityResponse.java
@@ -18,8 +18,8 @@ import gov.va.api.health.r4.api.elements.Extension;
 import gov.va.api.health.r4.api.elements.Meta;
 import gov.va.api.health.r4.api.elements.Narrative;
 import gov.va.api.health.r4.api.elements.Reference;
-import gov.va.api.health.r4.api.validation.ZeroOrOneOf;
-import gov.va.api.health.r4.api.validation.ZeroOrOneOfs;
+import gov.va.api.health.validation.api.ZeroOrOneOf;
+import gov.va.api.health.validation.api.ZeroOrOneOfs;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.Valid;
