+++ validation/pom.xml
@@ -1,16 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
   <parent>
+    <artifactId>api-starter</artifactId>
     <groupId>gov.va.api.health</groupId>
-    <artifactId>health-apis-parent</artifactId>
     <version>2.0.0</version>
   </parent>
-  <artifactId>fhir-resources-parent</artifactId>
+  <modelVersion>4.0.0</modelVersion>
+  <artifactId>validation</artifactId>
   <version>1.0.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-  <modules>
-    <module>r4</module>
-    <module>validation</module>
-  </modules>
+  <packaging>jar</packaging>
+  <dependencies>
+    <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-beans</artifactId>
+      <scope>compile</scope>
+    </dependency>
+  </dependencies>
 </project>
