+++ validation/src/main/java/gov/va/api/health/validation/api/ZeroOrOneOfValidator.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.r4.api.validation;
+package gov.va.api.health.validation.api;
 
 public class ZeroOrOneOfValidator extends AbstractSetFieldCounter<ZeroOrOneOf> {
